internal/telemetrytest: incorporated back into parent module
